+++ module/documents/_module.mjs
@@ -6,3 +6,5 @@ export * from "./actor.mjs";
 export * from "./item.mjs";
 export * from "./chat-message.mjs";
 export * from "./active-effect.mjs";
+export * from "./combat.mjs";
+export * from "./combatant.mjs";