+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable Spell Advances to mark their cost as additional to the advanced Spell's
 - Add input fields for system `identifier`s to sheets of relevant Items
 - Add labels and hoverable hints to armor and injury fields in edit mode adversary sheets
+- Implement initiative rolling for actors
 
 ### Changed